fix(core): Fix bug of partner admin cannot revoke token of mangaged org user
1. add error check based on code review

+++ app/modules/core/auth/auth.js
@@ -203,6 +203,9 @@
     }
 
     function revokeUserAuthTokens(userName, orgId) {
+      if (!_.isString(userName) || !_.isString(orgId)) {
+        return $q.reject('Invalid parameters passed');
+      }
       var revokeUrl = OAuthConfig.getOAuthRevokeUserTokenUrl() + $window.encodeURIComponent(userName) + '&orgid=' + orgId;
       return $http.delete(revokeUrl);
     }
